tbl2fase3/requirements.txt
django-tables2
django-filter==2.0.0
